useRef로 컴포넌트 내 변수 생성
