+++ tak-react/src/UserList.js
@@ -8,28 +8,11 @@ function User({ user }) {
   );
 }
 
-function UserList() {
-  const users = [
-    {
-      id: 1,
-      username: 'Wontak Lee',
-      email: 'takgung6820@gmail.com'
-    }, 
-    {
-      id: 2,
-      username: 'Company Tak',
-      email: 'wontak.lee@gwebscorp.com'
-    }, 
-    {
-      id: 3,
-      username: 'Bu-kae Tak',
-      email: 'roundt2505@naver.com'
-    }
-  ];
+function UserList({ users }) {
   return (
     <div>
       {users.map(user => (
-        <User user={user} />
+        <User user={user} key={user.id} />
       ))}
     </div>
   );
